LPD-88852 Rename OAuth2JWKValidator to OAuth2JWKValidatorUtil
