+++ modules/apps/oauth2-provider/oauth2-provider-api/bnd.bnd
@@ -7,7 +7,6 @@ Export-Package:\
 	com.liferay.oauth2.provider.exception,\
 	com.liferay.oauth2.provider.model,\
 	com.liferay.oauth2.provider.redirect,\
-	com.liferay.oauth2.provider.security.fips,\
 	com.liferay.oauth2.provider.service,\
 	com.liferay.oauth2.provider.service.persistence,\
 	com.liferay.oauth2.provider.util,\
