--- modules/apps/oauth2-provider/oauth2-provider-api/src/main/resources/com/liferay/oauth2/provider/security/fips/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
